Pull up common sink methods to abstract class
- These will be useful for implementing the puml sink
