+++ lib/brainstem/api_docs/sinks/open_api_specification_sink.rb
@@ -1,7 +1,6 @@
 require 'brainstem/api_docs'
 require 'brainstem/api_docs/formatters/open_api_specification/helper'
 require 'brainstem/api_docs/sinks/abstract_sink'
-require 'fileutils'
 require 'forwardable'
 
 module Brainstem
@@ -163,11 +162,7 @@ module Brainstem
         # Writes a given bufer to a filename within the base path.
         #
         def write_spec_to_file!
-          formatted_output = format_output(output, extension)
-          abs_path         = File.join(write_path, suggested_filename)
-
-          assert_directory_exists!(abs_path)
-          write_method.call(abs_path, formatted_output)
+          write_buffer_to_file(format_output(output, extension), suggested_filename)
         end
 
         #
@@ -180,30 +175,6 @@ module Brainstem
 
           data = output.to_hash
           requested_format == 'json' ? data.to_json : data.to_yaml
-        end
-
-        #
-        # Asserts that a directory exists, creating it if it does not.
-        #
-        def assert_directory_exists!(path)
-          dir = File.dirname(path)
-          FileUtils.mkdir_p(dir) unless File.directory?(dir)
-        end
-
-        #
-        # Defines how we write out the files.
-        #
-        def write_method
-          @write_method ||= Proc.new do |name, buff|
-            File.write(name, buff, mode: 'w')
-          end
-        end
-
-        #
-        # Defines the output directory for the file.
-        #
-        def write_path
-          @write_path ||= ::Brainstem::ApiDocs.write_path
         end
 
         #
